+++ libcaf_core/caf/blocking_actor.cpp
@@ -266,8 +266,13 @@ void blocking_actor::receive_impl(receive_cond& rcc, message_id mid,
       continue;
     }
     // Automatically unlink from actors after receiving an exit.
-    if (auto view = make_const_typed_message_view<exit_msg>(ptr->content()))
+    if (auto view = make_const_typed_message_view<exit_msg>(ptr->content())) {
       unlink_from(get<0>(view).source);
+    }
+    // Automatically clear incoming edges after receiving a down message.
+    if (auto view = make_const_typed_message_view<down_msg>(ptr->content())) {
+      clear_incoming_edges(get<0>(view).source);
+    }
     // Blocking actors can nest receives => push/pop `current_element_`
     auto prev_element = current_element_;
     current_element_ = ptr.get();